Implement DELETE method in login handler
An action has been added to handle 'DELETE' requests in the login handler of main.py. Previously, the DELETE method was not supported and the code would just pass without executing any action. Now, if a DELETE request is received, the Login Controller Handler will process the request and delete the login specified by the provided login_id.