Implement RoomUnavailable_Time_put method
The implementation of the RoomUnavailable_Time_put method in the roomunavailable.py file has been completed. This method retrieves the latest "enddate" from the "roomunavailable" table in the database where the "rid" matches and "ruid" doesn't match the given parameters. The connection is then closed after the execution of the query.