Add method to retrieve clients with most credit card reservations
A new method, Get_post_MostCreditCard, has been added to the LocalStatistics model. This method retrieves clients under the age of 30 who use a credit card and it limits the results to the top 5 clients with the highest number of reservations. The method groups results by client ID, age, payment method, hotel ID and room ID.