Add exception handling to login model DAO
The code for executing database queries in the login model DAO has been wrapped in a try/except block. This is to prevent the program from breaking if the execution of the query fails, by returning an "Error" message instead. This change has been applied to both the methods for inserting and updating login data.