Add 'MostReservation' method to GlobalStatistics model
A new function 'Get_post_MostReservation' has been added to the 'GlobalStatistics' model. This method retrieves the hotel's id and name, along with the total number of reservations for each hotel, then orders the results in descending order and limits the output based on 10% of the total number of hotels, which provides a leaderboard of hotels with the most reservations.