Update dictionary keys in LocalStatistics.py
The dictionary keys in the `HandicapRoom_Dict` and `LeastReserve_Dict` methods in `LocalStatistics.py` have been updated. A new `hid` key has been added and the index of the remaining keys has been altered accordingly. This change updates the structure of these dictionaries to better reflect the data they are representing.