Add method to get rooms with least guests
A new method `Get_post_LeastGuests` has been added to the `LocalStatistics` class in the `model_dao`. This method fetches the top 3 rooms with the least number of guests in them. It does so by executing a SQL query which groups and sorts rooms based on 'space' calculated as the difference in float values of guests and capacity.