Update SQL query in LocalStatistics.py
The SQL query used in the Get_post_RoomType function in the LocalStatistics.py file was revised. The 'hid' alias for 'hotel' was removed and a 'natural inner join' with the 'hotel' was added to the query. This adjustment improves the accuracy of the result set by ensuring the 'hotel' is directly linked with the other joined tables.