Update login functionality in dashboard notebook
The changes in the code mainly consist of an update to the login functionality in the dashboard.ipynb notebook. The initial straightforward login procedure is enhanced by adding sign-in functionality. Now, users can switch between login and sign-in views and effectively perform either operation. The way of displaying the login box and messages has also been improved to provide a better user experience.