Remove data length check in Check_Employee method
The length check for the data in the Check_Employee method of the employee controller has been removed. Also, the HTTP status code for a successful request has been explicitly set to 200. This means any number of elements in data will now be accepted and the status code will be more communicative of the request result.