Add cname to MostRevenue query and result dict
The code change refines the "MostRevenue" query and response dictionary in the GlobalStatistics.py (both in controller_handler and model_dao directories). The SQL query in the model_dao now includes 'cname' along with 'chid' in the GROUP BY clause. Consequently, 'cname' is also added to the resulting dictionary in the controller_handler to properly reflect these changes.