Add commented dbh.close after db.close in multiple files
The database handler close function, dbh.close(), has been added in a commented-out state in several files wherever the db.close() function was found. The intent behind this modification is to signify where dbh.close() could potentially be placed in the future if necessary. This change affects the model_dao directory files.