Dashboard Exoskeleton.
Login page welcomes the user. Log In and Sign in Buttons are created, the sign in button leads to the Create Account page. This page has a create button and a back button, which leads the user back to the log in interface.

Since this is still an early WIP, the page does not create or logs in the user.

The page also has a banner , icons in the buttons and a background, making it look fleshed out esthetically.