Added function to create 'room unavailable' in Jupyter notebook
The existing Jupyter notebook has been updated to include a new function that makes a room unavailable for a given period of time. This function, `Put_RoomUnavailable`, communicates with the backend via a Flask API. User interface has been provided to input the room ID and duration when the room will be unavailable.