Add MostCreditCard function and integrate into dashboard
Implemented the function MostCreditCard(id) in dashboard.ipynb for analysis of data based on credit card usage for reservations. This function will create a new section in the dashboard displaying the 'Top 5 clients under 30 that made most reservation with credit card' for the given hotel. It has been added to all user roles for easy access.